refactor(web): P5 — final carve wave; the composition root is 217 lines
RFC-003 P5, completing the carve. Fourteen sections move verbatim:
setup wizard, status & info, discord config, quick actions, personality,
self-update, startup diagnostics -> config_admin.py (+ self_update.py);
MCP, Slack, issue tracker, Grafana alerts, outbound webhooks ->
integrations.py; subsystem degradation appended to observability.py;
agent trajectories appended to sessions_chat.py (its true domain).

Two size-gate splits (RFC promised no module > ~600 lines): codex OAuth
-> codex_admin.py; self-update -> self_update.py. Largest module is now
llm_admin.py at 589.

src/web/api/__init__.py is 217 lines of pure composition: imports, the
ordered registrar call list the parity contract pins, setup_api, and
the R1 re-export surface. The 4,155-line monolith is fully dissolved
into 13 domain modules + api_common.

Composition assertions held; per-registrar admin gate added where
quick-actions used the root closure. Gate: new=0. Suite 6,079; parity
183 exact-order.
